unify src and include directories into RandLAPACK directory
* moved templated code from .cc files into .hh files. in the library is
  now header only
* added prefix to internal header file names
* added vritual destructors to polymorphic base classes
