+++ test/CMakeLists.txt
@@ -3,16 +3,16 @@ find_package(GTest)
 
 set(tmp FALSE)
 if (GTest_FOUND)
-
     set(tmp TRUE)
 
     set(RandLAPACK_test_srcs
-        src/comps/test_determiter.cc
-        src/comps/test_util.cc
-        src/comps/test_orth.cc
-        src/comps/test_qb.cc
-        src/drivers/test_rsvd.cc
-        src/drivers/test_cholqrcp.cc)
+        comps/test_determiter.cc
+        comps/test_util.cc
+        comps/test_orth.cc
+        comps/test_qb.cc
+        drivers/test_rsvd.cc
+        drivers/test_cholqrcp.cc
+    )
 
     add_executable(RandLAPACK_tests ${RandLAPACK_test_srcs})
